move object
Move also rename and copy between buckets

+++ caelus/aws/storages/s3_storage.py
@@ -67,21 +67,29 @@ class S3Storage(Storage):
         return self._list_s3_objects(self._get_folder_path(folder), filter_filename=filter_filename,
                                      only_files=only_files, filter_extension=filter_extension)
 
-    def _object_copy(self, dest_bucket_name: str, object_name: str, remove_copied: bool):
-        self.s3_resource.Object(dest_bucket_name, object_name).copy_from(CopySource={'Bucket': self.bucket_name,
-                                                                                     'Key': object_name})
-        self._aws_logger.debug(f'{object_name} copied from {self.bucket_name} to {dest_bucket_name}')
-        if remove_copied:
-            self.s3_resource.Object(self.bucket_name, object_name).delete()
-            self._aws_logger.debug(f'{object_name} removed from {self.bucket_name}')
+    def _object_copy(self, dest_bucket_name: str, object_name: str, dest_object_name: Union[str, None],
+                     remove_copied: bool):
+        if dest_object_name is None and dest_bucket_name == self.bucket_name:
+            self._aws_logger.warning(f'This config does not move the object')
+        else:
+            if dest_object_name is None and dest_bucket_name != self.bucket_name:
+                dest_object_name = object_name
 
-    def copy_between_storages(self, dest_name: str, files_to_move: Union[str, list, Generator],
-                              remove_copied: bool = False):
+            self.s3_resource.Object(dest_bucket_name, dest_object_name).copy_from(
+                CopySource={'Bucket': self.bucket_name,
+                            'Key': object_name})
+            self._aws_logger.debug(f'{object_name} copied from {self.bucket_name} to {dest_bucket_name}')
+            if remove_copied:
+                self.s3_resource.Object(self.bucket_name, object_name).delete()
+                self._aws_logger.debug(f'{object_name} removed from {self.bucket_name}')
+
+    def move_object(self, dest_storage_name: str, files_to_move: Union[str, list, Generator],
+                    dest_object_name: Union[str, None] = None, remove_copied: bool = False):
         if isinstance(files_to_move, str):
-            self._object_copy(dest_name, files_to_move, remove_copied)
+            self._object_copy(dest_storage_name, files_to_move, dest_object_name, remove_copied)
         else:
             for bucket_object in files_to_move:
-                self._object_copy(dest_name, bucket_object, remove_copied)
+                self._object_copy(dest_storage_name, bucket_object, dest_object_name, remove_copied)
 
     ###########
     # READERS #
